+++ src/bin/annatto.rs
@@ -1,12 +1,8 @@
-#[cfg(feature = "embed-documentation")]
-use annatto::documentation_server;
 use annatto::{
     error::AnnattoError,
     workflow::{execute_from_file, StatusMessage, Workflow},
     StepID,
 };
-#[cfg(feature = "embed-documentation")]
-use anyhow::anyhow;
 use indicatif::{ProgressBar, ProgressStyle};
 
 use std::{collections::HashMap, convert::TryFrom, path::PathBuf, sync::mpsc, thread};
@@ -30,9 +26,6 @@ enum Cli {
         #[structopt(parse(from_os_str))]
         workflow_file: std::path::PathBuf,
     },
-    #[cfg(feature = "embed-documentation")]
-    /// Show the documentation for this version of Annatto in the browser.
-    ShowDocumentation,
 }
 
 pub fn main() -> anyhow::Result<()> {
@@ -41,13 +34,6 @@ pub fn main() -> anyhow::Result<()> {
     let args = Cli::from_args();
     match args {
         Cli::Run { workflow_file, env } => convert(workflow_file, env)?,
-        #[cfg(feature = "embed-documentation")]
-        Cli::ShowDocumentation => {
-            let (_, _, handle) = documentation_server::start_server(true)?;
-            handle
-                .join()
-                .map_err(|_e| anyhow!("Waiting for documentation server thread failed"))??;
-        }
         Cli::Validate { workflow_file } => {
             Workflow::try_from((workflow_file, false))?;
         }
@@ -70,27 +56,34 @@ fn convert(workflow_file: PathBuf, read_env: bool) -> Result<(), AnnattoError> {
     let mut steps_progress: HashMap<StepID, f32> = HashMap::new();
 
     let bar = ProgressBar::new(1000);
-    bar.set_style(ProgressStyle::default_bar().template("[{elapsed}] [{bar:40}] {percent}% {msg}"));
+    bar.set_style(
+        ProgressStyle::default_bar()
+            .template("[{elapsed_precise}/est. {duration}] {wide_bar:.blue.bold} {percent}% {msg}")
+            .expect("Could not parse progress bar template"),
+    );
     let mut errors = Vec::new();
     for status_update in rx {
         match status_update {
             StatusMessage::Failed(e) => {
                 errors.push(e);
             }
+
             StatusMessage::StepsCreated(steps) => {
-                if steps.is_empty() {
-                    bar.println("No steps in workflow file")
-                } else {
-                    // Print all steps and insert empty progress for each step
-                    bar.println(format!("Conversion starts with {} steps", steps.len()));
-                    bar.println("-------------------------------");
-                    for s in steps {
-                        bar.println(format!("{}", &s));
-                        steps_progress.entry(s).or_default();
+                bar.suspend(|| {
+                    if steps.is_empty() {
+                        println!("No steps in workflow file")
+                    } else {
+                        // Print all steps and insert empty progress for each step
+                        println!("Conversion starts with {} steps", steps.len());
+                        println!("-------------------------------");
+                        for s in steps {
+                            println!("{}", &s);
+                            steps_progress.entry(s).or_default();
+                        }
+                        println!("-------------------------------");
                     }
-                    bar.println("-------------------------------");
-                }
-                bar.println("");
+                    println!();
+                });
             }
             StatusMessage::Info(msg) => {
                 bar.println(msg);