+++ src/importer/meta.rs
@@ -1,3 +1,7 @@
+//! Works similar to the Pepper configuration value
+//! [`pepper.before.readMeta`](https://corpus-tools.org/pepper/generalCustomizationProperties.html)
+//! and imports metadata property files for documents and corpora by using the file
+//! name as path to the document.
 use std::{
     collections::BTreeMap,
     io::{self, BufRead},